models/Chatbotbec/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-community
langchain-groq
requests